backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
asyncpg==0.29.0
databases[postgresql]==0.8.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
schedule==1.2.0
pytest==7.4.3
black==23.10.1
flake8==6.1.0
